+++ packages/fleur-react/CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.0.2
 - Fix incorrect typing in .d.ts
+- Improve connectToStores typing
 - Fix minor typing (internal)
 - Fix testing